create userTxn factory
