+++ hedera-node/hedera-app/src/test/java/com/hedera/node/app/workflows/handle/steps/UserTxnTest.java
@@ -20,7 +20,6 @@ import static com.hedera.hapi.node.base.HederaFunctionality.CONSENSUS_CREATE_TOP
 import static com.hedera.node.app.fixtures.AppTestBase.DEFAULT_CONFIG;
 import static com.hedera.node.app.service.token.impl.schemas.V0490TokenSchema.ACCOUNTS_KEY;
 import static com.hedera.node.app.workflows.handle.TransactionType.GENESIS_TRANSACTION;
-import static com.hedera.node.config.types.StreamMode.BLOCKS;
 import static java.util.Collections.emptyMap;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
@@ -37,11 +36,13 @@ import com.hedera.hapi.node.base.SignatureMap;
 import com.hedera.hapi.node.state.token.Account;
 import com.hedera.hapi.node.transaction.TransactionBody;
 import com.hedera.hapi.platform.event.EventTransaction;
+import com.hedera.node.app.blocks.BlockStreamManager;
 import com.hedera.node.app.blocks.impl.BoundaryStateChangeListener;
 import com.hedera.node.app.blocks.impl.KVStateChangeListener;
 import com.hedera.node.app.blocks.impl.PairedStreamBuilder;
 import com.hedera.node.app.fees.ExchangeRateManager;
 import com.hedera.node.app.fees.FeeManager;
+import com.hedera.node.app.records.BlockRecordManager;
 import com.hedera.node.app.service.consensus.impl.ConsensusServiceImpl;
 import com.hedera.node.app.service.token.api.FeeStreamBuilder;
 import com.hedera.node.app.services.ServiceScopeLookup;
@@ -154,6 +155,12 @@ class UserTxnTest {
     private NetworkUtilizationManager networkUtilizationManager;
 
     @Mock
+    private BlockRecordManager blockRecordManager;
+
+    @Mock
+    private BlockStreamManager blockStreamManager;
+
+    @Mock
     private WritableStates writableStates;
 
     @Mock
@@ -174,18 +181,9 @@ class UserTxnTest {
     void usesPairedStreamBuilderWithDefaultConfig() {
         given(configProvider.getConfiguration()).willReturn(new VersionedConfigImpl(DEFAULT_CONFIG, 1));
 
-        final var subject = UserTxn.from(
-                state,
-                event,
-                creatorInfo,
-                PLATFORM_TXN,
-                CONSENSUS_NOW,
-                GENESIS_TRANSACTION,
-                configProvider,
-                storeMetricsService,
-                kvStateChangeListener,
-                boundaryStateChangeListener,
-                preHandleWorkflow);
+        final var factory = createUserTxnFactory();
+        final var subject =
+                factory.createUserTxn(state, event, creatorInfo, PLATFORM_TXN, CONSENSUS_NOW, GENESIS_TRANSACTION);
 
         assertSame(GENESIS_TRANSACTION, subject.type());
         assertSame(CONSENSUS_CREATE_TOPIC, subject.functionality());
@@ -222,33 +220,11 @@ class UserTxnTest {
         given(accountState.getStateKey()).willReturn(ACCOUNTS_KEY);
         given(dispatcher.dispatchComputeFees(any())).willReturn(Fees.FREE);
 
-        final var subject = UserTxn.from(
-                state,
-                event,
-                creatorInfo,
-                PLATFORM_TXN,
-                CONSENSUS_NOW,
-                GENESIS_TRANSACTION,
-                configProvider,
-                storeMetricsService,
-                kvStateChangeListener,
-                boundaryStateChangeListener,
-                preHandleWorkflow);
+        final var factory = createUserTxnFactory();
+        final var subject =
+                factory.createUserTxn(state, event, creatorInfo, PLATFORM_TXN, CONSENSUS_NOW, GENESIS_TRANSACTION);
 
-        final var dispatch = subject.newDispatch(
-                authorizer,
-                networkInfo,
-                feeManager,
-                dispatchProcessor,
-                blockRecordInfo,
-                serviceScopeLookup,
-                storeMetricsService,
-                exchangeRateManager,
-                childDispatchFactory,
-                dispatcher,
-                networkUtilizationManager,
-                baseBuilder,
-                BLOCKS);
+        final var dispatch = factory.createDispatch(subject, baseBuilder);
 
         assertSame(PAYER_ID, dispatch.payerId());
         verify(baseBuilder).congestionMultiplier(CONGESTION_MULTIPLIER);
@@ -271,35 +247,33 @@ class UserTxnTest {
         given(accountState.getStateKey()).willReturn(ACCOUNTS_KEY);
         given(dispatcher.dispatchComputeFees(any())).willReturn(Fees.FREE);
 
-        final var subject = UserTxn.from(
-                state,
-                event,
-                creatorInfo,
-                PLATFORM_TXN,
-                CONSENSUS_NOW,
-                GENESIS_TRANSACTION,
+        final var factory = createUserTxnFactory();
+        final var subject =
+                factory.createUserTxn(state, event, creatorInfo, PLATFORM_TXN, CONSENSUS_NOW, GENESIS_TRANSACTION);
+
+        final var dispatch = factory.createDispatch(subject, baseBuilder);
+
+        assertSame(PAYER_ID, dispatch.payerId());
+        verify(baseBuilder, never()).congestionMultiplier(1);
+    }
+
+    private UserTxnFactory createUserTxnFactory() {
+        return new UserTxnFactory(
                 configProvider,
                 storeMetricsService,
                 kvStateChangeListener,
                 boundaryStateChangeListener,
-                preHandleWorkflow);
-
-        final var dispatch = subject.newDispatch(
+                preHandleWorkflow,
                 authorizer,
                 networkInfo,
                 feeManager,
                 dispatchProcessor,
-                blockRecordInfo,
                 serviceScopeLookup,
-                storeMetricsService,
                 exchangeRateManager,
-                childDispatchFactory,
                 dispatcher,
                 networkUtilizationManager,
-                baseBuilder,
-                BLOCKS);
-
-        assertSame(PAYER_ID, dispatch.payerId());
-        verify(baseBuilder, never()).congestionMultiplier(1);
+                blockRecordManager,
+                blockStreamManager,
+                childDispatchFactory);
     }
 }
